+++ src/v5/client.rs
@@ -162,6 +162,7 @@ where
         .build(factory(
             service.into_factory().map_err(MqttError::Service).map_init_err(MqttError::Service),
             self.control,
+            0,
         ))
         .map_err(|e| match e {
             DispatcherError::Service(e) => e,